refactor: Error 파일 생성 및 에러 Enum 위치 이동 #5
