+++ JuiceMaker/JuiceMaker/Model/FruitStore.swift
@@ -15,20 +15,6 @@ struct Fruit {
 
 class FruitStore {
     
-    enum FruitStoreError: LocalizedError {
-        case invalidFruitChoice
-        case lackOfStock(neededStock: Int)
-        
-        var description: String {
-            switch self {
-            case .invalidFruitChoice:
-                return "유효하지 않은 선택입니다."
-            case .lackOfStock(let neededStock):
-                return "재료가 \(neededStock)개 부족합니다. 재고를 확인해주세요."
-            }
-        }
-    }
-    
     private(set) var inventory: [Fruit] = []
  
     func initializeInventory() {
